feat(rts-core): stage 2b — migrate `path` namespace (Str proof) + Bool=i64
First string-heavy namespace on the single-declaration model, proving Str
support. path/mod.rs is now one impl block; abi.rs, components.rs, join.rs and
the orphan rt.rs are deleted (5 files -> 1). `Handle`-returning members use a
`ts = "...: string"` override (the handle is a GC string handle; derived TS
`number` can't know that).

- rts-abi/ty.rs: Bool repr fixed i8 -> i64 to match codegen's `AbiType::Bool`
  -> Cranelift I64 lowering (the old i8 was a latent width mismatch, benign for
  0/1). path::is_absolute and hint::assert_unchecked now match the calling
  convention exactly.
- jit.rs: path add_fn now reference the externs at the module root (glob
  `use path::*`) instead of the deleted `join::`/`components::` submodules.
- rts-codegen/abi SPECS: path::abi::SPEC -> path::SPEC.

Verified: cargo build --release green; rts.d.ts UNCHANGED (apis byte-identical);
runtime smoke ok — join, parent="a/b", ext="gz", is_absolute=false,
normalize("a/./b/../c")="a\c".

+++ crates/rts-abi/src/ty.rs
@@ -29,8 +29,10 @@ pub type I64 = i64;
 pub type I32 = i32;
 /// Native f64.
 pub type F64 = f64;
-/// Boolean at the extern "C" boundary (0/1 in an i8).
-pub type Bool = i8;
+/// Boolean at the extern "C" boundary (0/1). Codegen lowers `AbiType::Bool`
+/// to Cranelift `I64` for both params and returns, so the repr is `i64` — not
+/// `i8` — to match the calling convention exactly.
+pub type Bool = i64;
 
 /// Marker token for a UTF-8 string parameter. `#[rts_fn]` expands a `Str` arg
 /// into two ABI slots `(ptr: *const u8, len: i64)` and reconstructs an `&str`
